+++ src/main/java/com/atlassian/braid/document/DocumentMapperException.java
@@ -1,0 +1,9 @@
+package com.atlassian.braid.document;
+
+import static java.lang.String.format;
+
+public final class DocumentMapperException extends RuntimeException {
+    DocumentMapperException(String msg, Object... args) {
+        super(format(msg, args));
+    }
+}