+++ src/main/java/com/atlassian/braid/java/util/BraidMaps.java
@@ -1,5 +1,6 @@
 package com.atlassian.braid.java.util;
 
+import java.util.Iterator;
 import java.util.Map;
 import java.util.Optional;
 
@@ -14,5 +15,10 @@ public final class BraidMaps {
 
     public static <K, V> Optional<V> get(Map<K, V> map, K key) {
         return Optional.ofNullable(map.get(key));
+    }
+
+    public static <K, V> Map.Entry<K, V> firstEntry(Map<K, V> map) {
+        final Iterator<Map.Entry<K, V>> iterator = map.entrySet().iterator();
+        return iterator.hasNext() ? iterator.next() : null;
     }
 }