+++ src/main/java/com/atlassian/braid/mapper/Mapper.java
@@ -147,7 +147,7 @@ public interface Mapper extends UnaryOperator<Map<String, Object>> {
      * @param mapper a mapper for the new map
      * @return the mapper with the map operation, this is <em>not</em> necessarily the same mapper
      */
-    default Mapper map(String key, Mapper mapper) {
+    default Mapper map(String key, Function<Map<String, Object>, Map<String, Object>> mapper) {
         return map(key, __ -> true, mapper);
     }
 
@@ -159,7 +159,7 @@ public interface Mapper extends UnaryOperator<Map<String, Object>> {
      * @param mapper    a mapper for the new map
      * @return the mapper with the map operation, this is <em>not</em> necessarily the same mapper
      */
-    Mapper map(String key, Predicate<MapperInputOutput> predicate, Mapper mapper);
+    Mapper map(String key, Predicate<MapperInputOutput> predicate, Function<Map<String, Object>, Map<String, Object>> mapper);
 
     /**
      * Copies a map at a key to the same key in the target map