+++ src/main/java/com/atlassian/braid/source/GraphQLRemoteSchemaSource.java
@@ -4,15 +4,19 @@ import com.atlassian.braid.BraidContext;
 import com.atlassian.braid.Link;
 import com.atlassian.braid.SchemaNamespace;
 import com.atlassian.braid.SchemaSource;
+import com.atlassian.braid.document.DocumentMapper;
+import com.atlassian.braid.document.DocumentMappers;
 import graphql.ExecutionInput;
 import graphql.GraphQLError;
 import graphql.execution.DataFetcherResult;
+import graphql.schema.idl.TypeDefinitionRegistry;
 
 import java.io.Reader;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
@@ -28,7 +32,7 @@ import static java.util.Objects.requireNonNull;
 @SuppressWarnings("WeakerAccess")
 public final class GraphQLRemoteSchemaSource<C extends BraidContext> extends ForwardingSchemaSource<C> {
 
-    private final QueryExecutorSchemaSource<C> delegate;
+    private final BaseQueryExecutorSchemaSource<C> delegate;
     private final GraphQLRemoteRetriever<C> graphQLRemoteRetriever;
 
 
@@ -37,11 +41,21 @@ public final class GraphQLRemoteSchemaSource<C extends BraidContext> extends For
                                      GraphQLRemoteRetriever<C> graphQLRemoteRetriever,
                                      List<Link> links,
                                      String... topLevelFields) {
+        this(namespace, schemaProvider, graphQLRemoteRetriever, links, DocumentMappers.identity(), topLevelFields);
+    }
+
+    public GraphQLRemoteSchemaSource(SchemaNamespace namespace,
+                                     Supplier<Reader> schemaProvider,
+                                     GraphQLRemoteRetriever<C> graphQLRemoteRetriever,
+                                     List<Link> links,
+                                     Function<TypeDefinitionRegistry, DocumentMapper> documentMapper,
+                                     String... topLevelFields) {
         this.graphQLRemoteRetriever = requireNonNull(graphQLRemoteRetriever);
-        this.delegate = new QueryExecutorSchemaSource<>(namespace,
+        this.delegate = new BaseQueryExecutorSchemaSource<>(namespace,
                 loadPublicSchema(schemaProvider, topLevelFields),
                 loadSchema(schemaProvider),
                 links,
+                documentMapper,
                 this::query);
     }